updated few methods, moved FilterSpec to utils
