+++ src/main/resources/application.properties
@@ -1,4 +1,3 @@
-spring.application.name=
 server.port=8081
 
 ## Database
